moved gen_ply() to module
